# Conflicts:
#   backend/src/main/java/com/dollyanddot/kingmaker/domain/notification/service/NotificationServiceImpl.java

+++ backend/src/main/java/com/dollyanddot/kingmaker/domain/reward/repository/RewardRepository.java
@@ -3,5 +3,5 @@ package com.dollyanddot.kingmaker.domain.reward.repository;
 import com.dollyanddot.kingmaker.domain.reward.domain.Reward;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RewardRepository extends JpaRepository<Reward, Long> {
+public interface RewardRepository extends JpaRepository<Reward, Integer> {
 }
